fix(qwen4exp): restore full-width 310P QSA scores and coherent decode
Reduce all FP32 head dimensions in sparse attention, index scoring, and compressed-key normalization. Preserve PLE serving history, use exact dense attention within the QSA budget, and speed up 310P expert execution with W8A16 weight-only matmul. Add NPU regression tests for the previously truncated reductions.

+++ vllm_ascend/models/qwen4_exp/ops/qsa_indexer.py
@@ -81,11 +81,13 @@ def _stable_topk_indices(scores: torch.Tensor, k: int) -> torch.Tensor:
 
     # Stable least-significant-key sorts build the lexicographic order:
     # score descending, valid before padding, then group index ascending.
-    order = torch.argsort(candidate_indices, dim=1, stable=True)
+    # 310P sorts integer tensors on AiCPU. Group ids fit exactly in float32,
+    # so these two key sorts can stay on AiCore without changing tie order.
+    order = torch.argsort(candidate_indices.to(torch.float32), dim=1, stable=True)
     candidate_indices = candidate_indices.gather(1, order)
     candidate_values = candidate_values.gather(1, order)
     candidate_valid = candidate_valid.gather(1, order)
-    order = torch.argsort(candidate_valid.to(torch.int8), dim=1, descending=True, stable=True)
+    order = torch.argsort(candidate_valid.to(torch.float32), dim=1, descending=True, stable=True)
     candidate_indices = candidate_indices.gather(1, order)
     candidate_values = candidate_values.gather(1, order)
     order = torch.argsort(candidate_values, dim=1, descending=True, stable=True)
